Migrate runners unit tests to pytest
